Address PR review comments: - Changed `inboxId` type to long - Moved commented code to the integration test fixture

+++ src/lib/Mailtrap.Abstractions/Email/IEmailClient.cs
@@ -92,6 +92,6 @@ public interface IEmailClient
     Task<SendEmailResponse?> SendEmail(
         SendEmailRequest request,
         SendEndpoint? endpoint = default,
-        int? inboxId = default,
+        long? inboxId = default,
         CancellationToken cancellationToken = default);
 }
